backend/requirements.txt
asgiref==3.11.0
Django==5.2.10
djangorestframework==3.16.1
joblib==1.5.3
numpy==2.4.2
pandas==3.0.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
sqlparse==0.5.5
threadpoolctl==3.6.0
tzdata==2025.3
